+++ CMakeLists.txt
@@ -315,10 +315,10 @@ set(${PROJECT_NAME}_SOURCES
   src/nexus.h
   src/persistence/db/rawdatabase.cpp
   src/persistence/db/rawdatabase.h
+  src/persistence/db/upgrades/dbupgrader.cpp
+  src/persistence/db/upgrades/dbupgrader.h
   src/persistence/history.cpp
   src/persistence/history.h
-  src/persistence/dbupgrader.cpp
-  src/persistence/dbupgrader.h
   src/persistence/ifriendsettings.cpp
   src/persistence/ifriendsettings.h
   src/persistence/igroupsettings.cpp
